Changed Database Column
Changed the OrderedItems database column "menu_item_id" to "menu_item_name", so had to also change references to it in the files.

I changed the column name and data type because prior the column had a foreign key reference which did NOT allow menu items to be deleted from the MenuItem table due to the OrderedItems table referencing the primary key.